feat: removes handler in favor of simple method
